Rename ObserveLoginRequest.kt to FetchSiteData.kt
