+++ WooCommerce-Wear/src/test/java/com/woocommerce/android/ui/login/LoginViewModelTest.kt
@@ -5,8 +5,8 @@ import androidx.navigation.NavHostController
 import com.woocommerce.android.BaseUnitTest
 import com.woocommerce.android.phone.PhoneConnectionRepository
 import com.woocommerce.android.ui.NavRoutes
-import com.woocommerce.android.ui.login.ObserveLoginRequest.LoginRequestState.Logged
-import com.woocommerce.android.ui.login.ObserveLoginRequest.LoginRequestState.Timeout
+import com.woocommerce.android.ui.login.FetchSiteData.LoginRequestState.Logged
+import com.woocommerce.android.ui.login.FetchSiteData.LoginRequestState.Timeout
 import com.woocommerce.commons.wear.MessagePath.REQUEST_SITE
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.flowOf
@@ -25,14 +25,14 @@ import org.mockito.kotlin.whenever
 class LoginViewModelTest : BaseUnitTest() {
 
     private lateinit var sut: LoginViewModel
-    private val observeLoginRequest: ObserveLoginRequest = mock()
+    private val fetchSiteData: FetchSiteData = mock()
     private val phoneConnectionRepository: PhoneConnectionRepository = mock()
     private val navController: NavHostController = mock()
 
     @Test
     fun `when user is logged in, navigate to MY_STORE`() = testBlocking {
         // Given
-        whenever(observeLoginRequest.invoke()).thenReturn(flowOf(Logged))
+        whenever(fetchSiteData.invoke()).thenReturn(flowOf(Logged))
 
         // When
         createSut()
@@ -48,7 +48,7 @@ class LoginViewModelTest : BaseUnitTest() {
     fun `when user is not logged in, loading is stopped`() = testBlocking {
         // Given
         var isLoading: Boolean? = null
-        whenever(observeLoginRequest.invoke()).thenReturn(flowOf(Timeout))
+        whenever(fetchSiteData.invoke()).thenReturn(flowOf(Timeout))
         createSut()
         sut.viewState.observeForever { isLoading = it.isLoading }
 
@@ -99,7 +99,7 @@ class LoginViewModelTest : BaseUnitTest() {
 
         // Then
         verify(phoneConnectionRepository).sendMessage(REQUEST_SITE)
-        verify(observeLoginRequest).invoke()
+        verify(fetchSiteData).invoke()
     }
 
     @Test
@@ -115,14 +115,14 @@ class LoginViewModelTest : BaseUnitTest() {
 
         // Then
         verify(phoneConnectionRepository).sendMessage(REQUEST_SITE)
-        verify(observeLoginRequest, never()).invoke()
+        verify(fetchSiteData, never()).invoke()
         assertThat(isLoading).isNotNull()
         assertThat(isLoading).isFalse
     }
 
     private fun createSut() {
         sut = LoginViewModel(
-            observeLoginRequest,
+            fetchSiteData,
             phoneConnectionRepository,
             navController,
             SavedStateHandle()
